Added stuff related to microcode:    The microcode assembler    The microcode source    The opcode map    .gitignore to avoid versioning temporary files
As of this commit, the assembler writes to STDOUT including a lot of info about
the assembly process. To use, send the output to a file and delete the beginning of the
file up until the VHDL begins.  The result can be used to overwrite control_store.vhd.

Also added an empty "vivado_build" dir where project can be built without worrying about
adding unnecessary files to the src or from_wizard directories.